refactor: consistent positioning of connect wallet prompt
