+++ apps/main/src/components/PageCreatePool/Page.tsx
@@ -6,18 +6,21 @@ import { useLocation, useNavigate, useParams } from 'react-router-dom'
 import styled from 'styled-components'
 
 import { breakpoints } from '@/ui/utils/responsive'
+import { scrollToTop } from '@/utils'
+import usePageOnMount from '@/hooks/usePageOnMount'
+import useStore from '@/store/useStore'
 
 import DocumentHead from '@/layout/default/DocumentHead'
 import PoolCreation from '@/components/PageCreatePool/index'
-
-import { scrollToTop } from '@/utils'
-import usePageOnMount from '@/hooks/usePageOnMount'
+import Box from '@/ui/Box'
+import ConnectWallet from '@/components/ConnectWallet'
 
 const Page: NextPage = () => {
   const params = useParams()
   const location = useLocation()
   const navigate = useNavigate()
   const { curve } = usePageOnMount(params, location, navigate)
+  const provider = useStore((state) => state.wallet.getProvider(''))
 
   useEffect(() => {
     scrollToTop()
@@ -26,9 +29,21 @@ const Page: NextPage = () => {
   return (
     <>
       <DocumentHead title={t`Create Pool`} />
-      <Container>
-        <PoolCreation curve={curve as CurveApi} />
-      </Container>
+      {!provider ? (
+        <Box display="flex" fillWidth>
+          <ConnectWalletWrapper>
+            <ConnectWallet
+              description="Connect wallet to access pool creation"
+              connectText="Connect Wallet"
+              loadingText="Connecting"
+            />
+          </ConnectWalletWrapper>
+        </Box>
+      ) : (
+        <Container>
+          <PoolCreation curve={curve as CurveApi} />
+        </Container>
+      )}
     </>
   )
 }
@@ -45,6 +60,16 @@ const Container = styled.div`
   @media (min-width: ${breakpoints.lg}rem) {
     margin: 1.5rem;
   }
+`
+
+const ConnectWalletWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  justify-content: center;
+  margin: var(--spacing-5) auto auto;
+  background: var(--page--background-color);
+  padding: var(--spacing-4) var(--spacing-4);
 `
 
 export default Page
